Transform ratings from real raters into pseudo ratings

Data from large-scale assessments often are rated by multiple raters. This function reduces the number of raters by the use of “pseudo raters”.

Usage

make.pseudo (datLong, idCol, varCol, codCol, valueCol, n.pseudo, randomize.order = TRUE,
             verbose = FALSE)

Arguments

datLong

Data set in the longformat, i.e. one row per examinee-variable-rater combination.

idCol

Name or column number of the person identifier (ID) variable.

varCol

Name or column number of the variable identifier.

codCol

Name or column number of the rater identifier variable.

valueCol

Name or column number of the value variable.

n.pseudo

How many pseudo rater should be used? (value must be lower than the number of real raters)

randomize.order

Logical: if TRUE, the selection of raters to pseudo raters is random.

verbose

Logical: give information about number of persons, variables, raters?

Value

A data.frame in the long format.

Author

Sebastian Weirich

Examples

data(rater)
oneRater <- make.pseudo (datLong=rater, idCol="id", varCol="variable", codCol="rater",
            valueCol="value", n.pseudo=1)
twoRaters<- make.pseudo (datLong=rater, idCol="id", varCol="variable", codCol="rater",
            valueCol="value", n.pseudo=2)